requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124